+++ db/migrations/0006-20141210-user_activation.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `users` ADD `verification_key` VARCHAR( 255 ) NULL ;